fix: 중복된 알림은 재렌더링하지 않도록 컴포넌트 분리 및 useMemo 사용
